+++ controllers/player_controller.py
@@ -1,17 +1,17 @@
 from models.player_model import PlayerModel
 from views.cli_view import CLIView
-from models.language_manager import LanguageManager
+from models.language_model import LanguageManager
+
 
 class PlayerController:
     """Controller for managing player interactions."""
 
-    def __init__(self, 
-                 model: PlayerModel, 
-                 view: CLIView, 
-                 lang_manager: LanguageManager):
+    def __init__(
+        self, model: PlayerModel, view: CLIView, lang_manager: LanguageManager
+    ):
         """
         Initialize PlayerController.
-        
+
         Args:
             model (PlayerModel): Player data model
             view (CLIView): Command-line interface view
@@ -24,26 +24,26 @@ class PlayerController:
     def select_player(self) -> str:
         """
         Interactive player selection process.
-        
+
         Returns:
             Selected player name
         """
         players = self._model.get_players()
-        
+
         while True:
             # Display players
             self._view.display_menu(
                 title=self._lang.get("player_management.select_player"),
-                options=players + [self._lang.get("player_management.add_new_player")]
+                options=players + [self._lang.get("player_management.add_new_player")],
             )
-            
+
             # Get user choice
             choice = self._view.get_numeric_input(
                 prompt=self._lang.get("player_management.enter_choice"),
-                min_value=1, 
-                max_value=len(players) + 1
+                min_value=1,
+                max_value=len(players) + 1,
             )
-            
+
             # Handle selection
             if choice <= len(players):
                 selected_player = self._model.select_player(choice)
@@ -53,12 +53,12 @@ class PlayerController:
                     )
                 )
                 return selected_player
-            
+
             # Add new player
             new_player = self._view.get_text_input(
                 prompt=self._lang.get("player_management.enter_player_name")
             )
-            
+
             added_player = self._model.add_player(new_player)
             if added_player:
                 self._view.display_message(
