+++ legacy/level_management.py
@@ -1,4 +1,4 @@
-from models.language_manager import LanguageManager
+from models.language_model import LanguageManager
 
 
 def select_level(lang_manager: LanguageManager):
