+++ models/__init__.py
@@ -3,7 +3,7 @@ from utilities.logging_util import logger
 logger.info("Models module initialized.")
 
 # Import models
-from .vocabulary import Vocabulary
+from .vocabulary_model import Vocabulary
 
 # from .mastery import Mastery
 # from .score import Score
