+++ models/vocabulary_model.py
